refactor(game): domain과 view를 classicLotto로 이동
확장 가능한 구조를 위해 domain과 view를 classicLotto 패키지로 이동
- domain, view 패키지 변경
- 모든 파일의 import 경로 수정
