+++ src/main/java/lotto/game/classicLotto/domain/LottoResult.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.game.classicLotto.domain;
 
 import java.util.Arrays;
 import java.util.EnumMap;
