+++ src/main/java/lotto/game/classicLotto/view/dto/RankStatistic.java
@@ -1,4 +1,4 @@
-package lotto.view.dto;
+package lotto.game.classicLotto.view.dto;
 
 public record RankStatistic(
         String description,
